Split semantic feature staging into focused Rust modules
