+++ bitloops/src/capability_packs/semantic_clones/stage_semantic_features/tests.rs
@@ -1,0 +1,3 @@
+mod current_projection;
+mod summary;
+mod support;
